fix:budgetAmount assigned value

+++ js/script.js
@@ -43,7 +43,7 @@ document.addEventListener("DOMContentLoaded", function () {
    document.querySelector('.add-budget-container form').addEventListener('submit', function(event){
     event.preventDefault();
     let budgetInput = document.getElementById('budget');
-    let budgetAmount = parseInt(budgetInput.ariaValueMax.trim());
+    let budgetAmount = parseFloat(budgetInput.ariaValueMax.trim());
 
     if(isNaN(budgetAmount) || budgetAmount <= 0) {
         alert('Please enter a valid budget amount!')
